# investigate loop unrolling 