Recently I've learned that all Intel processors since Sandy Bridge can do two `_mm_loadu_si128` at the same time with port 2 and 3. So I tried 2 sequential `_mm_loadu_si128` and it was a success. Then I also tried 4 and 8. 4 gave me an additional boost, but not 8.
Alas, when I pulled upstream commits, I got a significant performance penalty for the `esperanto` file with `msvc`. So I dropped them and started adding one by one. And I found which one causes it.
7761599262953df2a1d9c3427d0d27d1cb044615 SSE UTF16 => latin1 (#311)
It seems there's nothing special here. It just added a new dependency with 2 other `sse` implementations.
So I also checked with `gcc` and there was no penalty.
Could it be a `msvc` bug?

`gcc` 11.4.0 under `wsl2`
`msvc` 17.5.5
|                                     | esperanto    | french              |  german    | portuguese  |
| -------------                 | -------------   | -------------   | -------------   | ------------- |
|           `msvc`                  |                      |                        |                        |                    |
|   master                       | 12.641 GB/s  | 3.714 GB/s      | 5.845 GB/s    | 4.246 GB/s   |
|                                     | 100%            |   100%             |  100%            |  100%          |
| before the commit      | 16.111 GB/s  | 4.638 GB/s      | 8.743 GB/s    | 5.972 GB/s    |
|                                     |  127.45%        |   124.88%      |  149.58%       |  140.65%        |
| after  the commit       | 12.450 GB/s  | 4.522 GB/s      | 8.237 GB/s     | 5.869 GB/s   |
|                                     |  98.49%         |   121.76%      |  140.92%       |  138.22%        |
|           `gcc`                    |                       |                        |                        |                    |
| master                        | 12.839 GB/s   | 3.717 GB/s     | 6.921 GB/s      | 4.583 GB/s   |
|                                     | 100%            |   100%             |  100%            |  100%          |
| before the commit    | 14.673 GB/s    | 4.384 GB/s     | 8.704 GB/s      | 5.920 GB/s   |
|                                    | 114.28%         |   117.94%      |  125.76%         |  129.17%     |
| after  the commit      | 14.673 GB/s   | 4.389 GB/s     | 8.667 GB/s      | 5.920 GB/s   |
|                                    | 114.28%         |   118.08%      |  125.23%         |  129.17%     |


The situation got even funnier when I removed all the loops except [this one](https://github.com/simdutf/simdutf/compare/master...aspic-fish:simdutf:sse_convert_latin1_to_utf8_perf#diff-2032dccf513eb16b21de09f8f901e889a0042a2e6e894a0d03a3f1766161ebf8R86), and got the opposite result. And it's quite consistent between benchmarks.
|                                     | esperanto    | french              |  german    | portuguese  |
| -------------                 | -------------   | -------------   | -------------   | ------------- |
|           msvc                  |                      |                        |                        |                    |
| before the commit      | 14.673 GB/s  | 4.102 GB/s      | 7.144 GB/s     | 4.896 GB/s    |
| after  the commit        | 15.503 GB/s   |  4.205 GB/s     | 7.608 GB/s     | 5.108 GB/s   |

I'm going to continue the investigation in a couple of days.
plan: 
* I suspect that building it as a shared lib might help as it would prevent access of `msvc` to the rest of the code.
Supposedly, that wouldn't allow it to perform some smart optimisations and thus results should be more stable. 
* try more compilers.
* check how adding/removing other `sse` implementations affects performance
* try unrolling other implementations as well


For now, I suggest considering unrolling as unstable. Any ideas?