@aqrit 2023-09-20
Could you add some clarity?
The PR is for `sse_convert_latin1_to_utf8`.
But you're discussing a performance regression with `sse_convert_utf16_to_latin1` ?

This PR :
I suspect the branch for 'ASCII fast path' will interfere with unrolling attempts.

The bottleneck is probably: latency from the random lookups and the thruput of the shuffle port.
(I imagine there are 6 shuffles per 16 bytes of input) 

The CPU can have 20+ loads "in-flight" at once. The total loads issued per cycle is not that interesting for unrolling (we always have the same number of loads).  It is mostly for scheduling instructions at the assembly level, and to ballpark whether we're IO bound (e.g. do we want to replace a lookup with a calculation).