@aspic-fish 2023-09-20
Tables contain results for `sse_convert_latin1_to_utf8` only. But somehow 7761599262953df2a1d9c3427d0d27d1cb044615 commit causes `sse_convert_latin1_to_utf8` regression.
I just go to the brunch [sse_convert_latin1_to_utf8_perf](https://github.com/aspic-fish/simdutf/tree/sse_convert_latin1_to_utf8_perf), rebuild all and run the benchmark. Results are placed in the  "before the commit" row. 
Then I just do `git rebase 7761599262953df2a1d9c3427d0d27d1cb044615`, rebuild all and run the benchmark again. Results are placed in the "after the commit" row.