@aspic-fish 2023-09-21
@aqrit,
> The bottleneck is probably: latency from the random lookups and the thruput of the shuffle port.
(I imagine there are 6 shuffles per 16 bytes of input)

according to [uops](https://uops.info/table.html) from `Nehalem` to `Ivy Bridge` there are 2 ports for shuffle. 
And I use 2 `_mm_shuffle_epi8` per `_mm_load_si128` to split input into 2 vectors. They are pipelined and should go in 1 cycle both.
And 2 more to pack vectors into `utf-8` just after lookup. +1 cycle per shuffle
I don't think shuffle is a candidate for a bottleneck. But `LUT` surely is. Google says `L1` cache hit takes 4 cycles in best case.
There're 2 lookups per load, so at least 8 cycles or many more in case of a cache miss.

But  `_mm_load_si128` is a bottleneck too. A smaller one. It's latency is 6.  My original thought was that it actually might not be pipelined automatically.  In this case, 2 sequential calls of it would give 32 bytes instead of 16 for the same latency.

> The CPU can have 20+ loads "in-flight" at once.

but `x86-64` has only 16 `sse` registers, how could it be 20+?
Could you provide keywords for googling?

PS: I'm actually new to all these simd stuff so excuse me if I say something stupid)