@lemire 2023-09-21
> x86-64 has only 16 sse registers, how could it be 20+?

These are named registers, but we have many more registers.

You can examine the issue experimentally...
https://lemire.me/blog/2022/06/07/memory-level-parallelism-intel-ice-lake-versus-amazon-graviton-3/