@aqrit 2023-09-21
A `Nehalem` would not have a problem with the shuffles. However, a `Haswell` or `Skylake` might. On what CPU are you performing the benchmark?

The terms would be [register renaming](https://en.wikipedia.org/wiki/Register_renaming) and [out-of-order execution](https://en.wikipedia.org/wiki/Out-of-order_execution#Out-of-order_processors), I guess.

[uops.info Code Analyzer](https://github.com/andreas-abel/uiCA) is probably a good place to start, if we want to micro-optimize this.