@lemire 2023-09-21
@aqrit A fun one is this PR: https://github.com/simdutf/simdutf/pull/318

The westmere kernel (which is currently just scalar code, but subject to autovectorization) is faster than a reasonable hand-coded AVX2 routine. It is still fine because the differences are small... but it is clear that we could micro-optimize better.