@aspic-fish 2023-09-23
I think I get it roughly. But I struggle to draw parallels with actual code in general. But fortunately, for this PR I'm not into tuning this exact implementation but rather introducing the `msvc` issue and whether it's present in other compilers as well as unrolling effect with different compilers.
 
I additionally checked `clang` and `icx`, both doesn't have this issue at least for this pr. But the results also showed that unrolling is not necessarily consistent between compilers for the same os. `clang` showed performance degradation for windows and ubuntu, unlike the rest. So, if we use unrolling, we have to choose how to handle such situations.

Tomorrow I'll try to found out whether building separately helps with `msvc`.
P.S. I got a build error with `mingw`, should I open an issue?