@aspic-fish 2023-09-24
It doesn't. But I found that `clang` can actually benefit from unrolling. Just don't use inlined functions in loop body. For some reason they drop performance a lot.