@aqrit 2023-10-01
> cache miss

If we're willing to do 4 lookups per 16 bytes of input, then we'd only use 2 cache lines for tables.
https://gist.github.com/aqrit/5c914da98006874d0401983eb687e30e 

Note: I haven't actually studied the utf16->utf8 function so I don't what that is doing...